Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart
email-validator
paho-mqtt
websockets
python-decouple
alembic
